+++ app/src/androidTest/java/com/lastaoutdoor/lasta/ui/moreinfo/MoreInfoScreenTest.kt
@@ -4,6 +4,7 @@ import androidx.activity.compose.setContent
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
+import androidx.navigation.compose.rememberNavController
 import com.lastaoutdoor.lasta.data.model.activity.ActivityType
 import com.lastaoutdoor.lasta.data.model.activity.OutdoorActivity
 import com.lastaoutdoor.lasta.di.AppModule
@@ -28,7 +29,10 @@ class ActivityScreenTest {
   fun setUp() {
     hiltRule.inject()
     val fakeActivity = OutdoorActivity(ActivityType.CLIMBING, 1, 1.5f, "3 hours", "Test Title")
-    composeRule.activity.setContent { ActivityScreen(activity = fakeActivity) }
+    composeRule.activity.setContent {
+      val navController = rememberNavController()
+      MoreInfoScreen(activity = fakeActivity, navController)
+    }
   }
 
   @Test
