+++ app/src/main/java/com/lastaoutdoor/lasta/ui/navigation/MenuNavGraph.kt
@@ -6,9 +6,11 @@ import androidx.compose.ui.platform.testTag
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import com.lastaoutdoor.lasta.data.model.activity.ActivityType
+import com.lastaoutdoor.lasta.data.model.activity.OutdoorActivity
+import com.lastaoutdoor.lasta.ui.screen.activities.MoreInfoScreen
 import com.lastaoutdoor.lasta.ui.screen.discovery.DiscoveryScreen
 import com.lastaoutdoor.lasta.ui.screen.favorites.FavoritesScreen
-import com.lastaoutdoor.lasta.ui.screen.moreinfo.MoreInfoScreen
 import com.lastaoutdoor.lasta.ui.screen.profile.ProfileScreen
 import com.lastaoutdoor.lasta.ui.screen.social.SocialScreen
 
@@ -18,17 +20,20 @@ fun MenuNavGraph(
     navController: NavHostController,
     modifier: Modifier
 ) {
+  val fakeActivity = OutdoorActivity(ActivityType.CLIMBING, 0, 1.5f, "3 hours", "Test Title")
   NavHost(
       navController = navController,
       route = RootScreen.Main.route,
       modifier = modifier.testTag("MenuNavGraph"),
       startDestination = LeafScreen.Discover.route) {
-        composable(LeafScreen.Discover.route) { DiscoveryScreen() }
+        composable(LeafScreen.Discover.route) { DiscoveryScreen(navController) }
         composable(LeafScreen.Favorites.route) { FavoritesScreen(navController) }
         composable(LeafScreen.Social.route) { SocialScreen() }
         composable(LeafScreen.Profile.route) {
           ProfileScreen(rootNavController = rootNavController)
         }
-        composable(LeafScreen.MoreInfo.route) { MoreInfoScreen(navController = navController) }
+        composable(LeafScreen.MoreInfo.route) {
+          MoreInfoScreen(activity = fakeActivity, navController)
+        }
       }
 }
